rendering plant info
default plant Information is now rendered on home page users can search plants and get info

*small error when trying to load img data
