+++ plantapp/src/index.js
@@ -5,11 +5,5 @@ import App from './App';
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-  
-);
+root.render( <App />);
