+++ plantapp/src/pages/Home.jsx
@@ -8,9 +8,7 @@ import { Link } from "react-router-dom";
 function Home() {
   return (
     <>
-    <div className="fixed z-20 flex w-full items-center">
-      <img className='w-full'src="/images/hangin-leaves.png" alt="" />
-    </div>
+    
     <div>
       <h1 className="text-3xl py-5 font-bold underline text-green-700">
         Super Duper Plant App!
@@ -26,8 +24,11 @@ function Home() {
       </button>
       </Link>
       </p>
-      <SearchBar placeholder={'Search for your favorite plants!'}/>
-    </div>
+      <div>
+      <SearchBar className ="display flex-auto " placeholder={'Search for your favorite plants!'}/>
+
+</div>
+        </div>
     </>
   )
 }
